server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
typing-extensions==4.8.0
